More 'MoreLikeThis'-Component refactoring
