+++ library/Solarium/Component/MoreLikeThis/Select/ResultCollection.php
