+++ tests/Solarium/Tests/Component/MoreLikeThis/Select/ResultCollectionTest.php
