+++ tests/Solarium/Tests/Component/MoreLikeThis/Select/ResultTest.php
